Fix properties synchronize issue in action (JSB)
