+++ cocos2d/actions/CCActionInstant.js
@@ -251,7 +251,7 @@ cc.FlipX = cc.ActionInstant.extend({
     },
 
     update:function (dt) {
-        this.target.flippedX = this._flippedX;
+        this.target.scaleX = Math.abs(this.target.scaleX) * (this._flippedX ? 1 : -1);
     },
 
     reverse:function () {
@@ -309,7 +309,7 @@ cc.FlipY = cc.ActionInstant.extend({
     },
 
     update:function (dt) {
-        this.target.flippedY = this._flippedY;
+        this.target.scaleY = Math.abs(this.target.scaleY) * (this._flippedX ? 1 : -1);
     },
 
     reverse:function () {
